refactor(triage): extract prompts and improve triage logic
- Extract prompt templates to prompts/ directory
- Remove embedded prompt strings from triage.py
- Simplify triage categorization logic
- Improve genai_handler for clarity
- Add prompt management to __init__
- Enhance code maintainability and prompt versioning
